Description
3D ping pong game is a game designed using Unity 3D software, Sensors and Arduino.
The user interface for the game was developed using unity 3d and with help of MPU6050 sensor which consisted of 3 axis accelerometer, the angle of tilt or inclination along the X Y and Z axes were detected. With the help of serial communication, connection was established between the sensor and the Arduino. The game consists of two levels and the sliders are assigned to each player , the player can move the slider by moving his/her wrist as the sensor were attached on the upper side of palms. To make a goal the player had to push the ping pong to the opponents boundary. In level 2 there was an obstacle placed in between so that the ping pong ball would collide and move randomly in the play area.